refactor: remove requestStorage, change of layout
