+++ src/main/java/ru/practicum/shareit/booking/Booking.java
@@ -2,7 +2,7 @@ package ru.practicum.shareit.booking;
 
 import lombok.Data;
 import ru.practicum.shareit.item.model.Item;
-import ru.practicum.shareit.user.User;
+import ru.practicum.shareit.user.model.User;
 
 import java.time.LocalDateTime;
