--- src/main/java/ru/practicum/shareit/request/ItemRequest.java
@@ -1,14 +1,0 @@
-package ru.practicum.shareit.request;
-
-import lombok.Data;
-import ru.practicum.shareit.user.User;
-
-import java.time.LocalDateTime;
-
-@Data
-public class ItemRequest {
-    private long id;
-    private String description;
-    private User requestor;
-    private LocalDateTime created;
-}
