+++ src/main/java/ru/practicum/shareit/request/model/ItemRequest.java
@@ -1,0 +1,22 @@
+package ru.practicum.shareit.request.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.*;
+import java.time.LocalDateTime;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@Entity
+@Table(name = "requests")
+public class ItemRequest {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private long id;
+    private String description;
+    private long requestorId;
+    private LocalDateTime created;
+}
