+++ src/main/java/ru/practicum/shareit/user/UserRepository.java
@@ -1,6 +1,7 @@
 package ru.practicum.shareit.user;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import ru.practicum.shareit.user.model.User;
 
 public interface UserRepository extends JpaRepository<User, Long> {
 }
