+++ src/main/java/ru/practicum/shareit/user/dto/UserMapper.java
@@ -1,6 +1,6 @@
 package ru.practicum.shareit.user.dto;
 
-import ru.practicum.shareit.user.User;
+import ru.practicum.shareit.user.model.User;
 
 
 public class UserMapper {
