+++ src/main/resources/schema.sql
@@ -4,10 +4,10 @@ CREATE TABLE IF NOT EXISTS users (
     email VARCHAR(512) UNIQUE NOT NULL
 );
 
-CREATE TABLE IF NOT EXISTS item_requests (
+CREATE TABLE IF NOT EXISTS requests (
     id BIGINT GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
     description VARCHAR(512),
-    user_id BIGINT REFERENCES users(id),
+    requestor_id BIGINT REFERENCES users(id),
     created TIMESTAMP
 );
 
@@ -17,7 +17,7 @@ CREATE TABLE IF NOT EXISTS items (
     description VARCHAR(512),
     available BOOLEAN,
     owner_id BIGINT REFERENCES users(id),
-    request_id BIGINT REFERENCES item_requests(id)
+    request_id BIGINT REFERENCES requests(id)
 );
 
 CREATE TABLE IF NOT EXISTS bookings (
